% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COST354.R
\name{get_COST354_index}
\alias{get_COST354_index}
\title{Calculate a Health Index using the COST354 Report Method}
\usage{
get_COST354_index(index_values, weights, max_allowed, influence_fact = 0.4)
}
\arguments{
\item{index_values}{vector containing the values for the various sub-criteria.
These are assumed to be normalised to the same scale as per COST354
methodology}

\item{weights}{Relative Importance Weights for the various indexes to be
aggregated. Weights should be between 0 and 1, with zero denoting low
importance and 1 denoting highest importance. Weights do NOT have to add
up to 1 (see COST354 report pages 34 to 36 for examples)}

\item{max_allowed}{Maximum return value allowed. If the calculation results
in a number greater than this, then \code{max_allowed} is returned.}

\item{influence_fact}{'influence factor' as per COST354 methodology. This
value is typically in the range of 0.3 to 0.5. See COST354 report for details.}
}
\value{
Index value that aggregates the index values using COST354 method
}
\description{
\code{get_COST354_index} Calculates a normalised, weighted index based on the
methodology in the European COST354 report.
}
