% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_methods1.R
\name{jc_check_req_cols}
\alias{jc_check_req_cols}
\title{Checks for a list of Required Columns in a data frame}
\usage{
jc_check_req_cols(data, required_cols, loud = TRUE)
}
\arguments{
\item{data}{Data frame with source data on which the columns will be checked}

\item{required_cols}{Character vector containing the list of required columns}

\item{loud}{If TRUE, warning messagea are shown listing missing columns, '
if any}
}
\value{
TRUE if all required columns are present, otherwise FALSE
}
\description{
\code{jc_check_req_cols} checks a list of columns to see if all columns are
present in a source data set
}
