% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_methods1.R
\name{jc_clamp_columns}
\alias{jc_clamp_columns}
\title{Clamps values in numeric columns to specified minimum and maximum values}
\usage{
jc_clamp_columns(data, min_max_values)
}
\arguments{
\item{data}{Data frame with source data in which the columns will be clamped}

\item{min_max_values}{Data frame containing information on the columns to
be clampled. This dataframe must contain column names: 'column_name',
'min_allowed' and 'max_allowed' (case sensitive) with matching values mapping
to the column name in the source data (assumed to be a numeric column).}
}
\value{
A data frame identical to the source data but with values clamped to
the minimums and maximums for the specified columns
}
\description{
\code{jc_clamp_columns} Loops over a set of specified columns, each with an
accompanying minimum and maximum allowed. For each column, any values outside
the specified minimum or maximum are clamped to the minimum or maximum,
respectively.
}
\details{
This method will automatically convert all of the columns listed in
the dataframe \code{min_max_values} to numeric type, and any text values
found will be converted to NA.
}
