% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_helpers.R
\name{jc_do_multi_col_lookups}
\alias{jc_do_multi_col_lookups}
\title{Assigns data within an existing or new column based on a lookup set}
\usage{
jc_do_multi_col_lookups(data, lookup_columns, lookup_sets)
}
\arguments{
\item{data}{Data frame with raw input data in which the columns will
be converted}

\item{lookup_columns}{Data frame containing information on the columns to
be looked up. This dataframe must contain column names: 'source_column',
'lookup_name' and 'target_column'. Here, source_column must map to a name in
the raw input data, and lookup_name must map to one of the lookup names
specified in the lookup_sets parameter}

\item{lookup_sets}{Data frame containing all the lookup sets to be used for
the various columns.This dataframe must contain column names: 'lookup_name',
'keys' and 'values'. Keys specifies the allowed values that will be looked
up on, and values specifies the values that will be assigned.}
}
\value{
A data frame with the looked-up values
}
\description{
\code{jc_do_multi_col_lookups} does character data assignment in multiple
columns based on lookup sets specified for each column. The looked-up values
can be assigned into an new column or into an in-place column.
}
\details{
For example, an existing column may contain long descriptive names and you
want to convert these to shorter codes. A lookup table for that column is
then used to either re-assign the values in the column (in this case the
'source column' is the same as the 'target column'), or to add a new
column ('target column' does NOT match 'source column') that will contain the
shorter codes.

This method essentially mimics a repeated VLOOKUP() in Excel, that is
performed over a set of specified columns.

NOTE: Any values in the source column that are not found in the lookup set for the
column will be converted to NA
}
