% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_breakdown_stats.R
\name{jc_get_breakdown_length_pct}
\alias{jc_get_breakdown_length_pct}
\title{Gets a breakdown of percentage length for a specific parameter over epochs.}
\usage{
jc_get_breakdown_length_pct(
  param_data,
  epochs,
  bins,
  network_length,
  loc_from_col = "loc_from",
  loc_to_col = "loc_to"
)
}
\arguments{
\item{param_data}{Forecasted data for the parameter, as loaded from the export
file created by jcass.}

\item{epochs}{Vector with the calendar epochs for which to get breakdown The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{bins}{Bins with the limits for different breakdown groups.}

\item{network_length}{Total network length on which to base percentages.
Assumed to be in same unit as the location from and to column data.}

\item{loc_from_col}{Name of the column with the From/Start locations.}

\item{loc_to_col}{Name of the column with the To/End locations.}
}
\description{
\code{jc_get_breakdown_length_pct} Gets a table with the breakdown of the
percentage of network length in each bin for predicted parameter data for a
specific parameter over epochs.
}
