% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_breakdown_plots.R
\name{jc_get_breakdown_plot}
\alias{jc_get_breakdown_plot}
\title{Gets a breakdown plot of values for a specific parameter over epochs.}
\usage{
jc_get_breakdown_plot(
  param_data,
  epochs,
  bins,
  colours = NULL,
  legend_title = "Category"
)
}
\arguments{
\item{param_data}{Forecasted data for the parameter, as loaded from the export
file created by jcass.}

\item{epochs}{Vector with the calendar epochs for which to get breakdown The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{bins}{Bins with the limits for different breakdown groups.}

\item{colours}{Optional vector with colours to match each bin. Number should
be at least one less than the size of the bins vector. If the default value
of NULL is passed, then jcassr will automatically choose stoplight colours.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_breakdown_plot} Gets a plot with the frequency breakdown of
predicted parameter data for a specific parameter over epochs.
}
