% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_breakdown_plots.R
\name{jc_get_breakdown_plot_pct_len}
\alias{jc_get_breakdown_plot_pct_len}
\title{Gets a breakdown plot of percentage length for a specific parameter over epochs.}
\usage{
jc_get_breakdown_plot_pct_len(
  param_data,
  epochs,
  bins,
  network_length,
  loc_from_col = "loc_from",
  loc_to_col = "loc_to",
  colours = NULL,
  legend_title = "Category"
)
}
\arguments{
\item{param_data}{Forecasted data for the parameter, as loaded from the export
file created by jcass.}

\item{epochs}{Vector with the calendar epochs for which to get breakdown The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{bins}{Bins with the limits for different breakdown groups.}

\item{network_length}{Total network length on which to base percentages.
Assumed to be in same unit as the location from and to column data.}

\item{loc_from_col}{Name of the column with the From/Start locations.}

\item{loc_to_col}{Name of the column with the To/End locations.}

\item{colours}{Optional vector with colours to match each bin. Number should
be at least one less than the size of the bins vector. If the default value
of NULL is passed, then jcassr will automatically choose stoplight colours.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_breakdown_plot_pct_len} Gets a plot with the percentage length
breakdown of predicted parameter data for a specific parameter over epochs.
}
