% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_breakdown_stats.R
\name{jc_get_breakdown_table}
\alias{jc_get_breakdown_table}
\title{Gets a breakdown of values for a specific parameter over epochs.}
\usage{
jc_get_breakdown_table(param_data, epochs, bins)
}
\arguments{
\item{param_data}{Forecasted data for the parameter, as loaded from the export
file created by jcass.}

\item{epochs}{Vector with the calendar epochs for which to get breakdown The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{bins}{Bins with the limits for different breakdown groups.}
}
\description{
\code{jc_get_breakdown_table} Gets a table with the frequency breakdown of
predicted parameter data for a specific parameter over epochs. Here, the
frequency is the number of elements with condition in each bin.
}
