% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graphics.R
\name{jc_get_colours}
\alias{jc_get_colours}
\title{Gets a vector of colour codes from green to red}
\usage{
jc_get_colours(nbins)
}
\arguments{
\item{nbins}{number of bins to match the colours to. Valid values are 3,4 and 6}
}
\value{
Vector of HTML colour codes
}
\description{
\code{jc_get_colours} gets a set of colours representing a green-to-red
spectrum, based on a given number of matching bins. Order of colours
returned is green first, then yellow, then on to red etc.
}
