% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_cross_cut.R
\name{jc_get_crosscut_graph}
\alias{jc_get_crosscut_graph}
\title{Gets a graph showing a Cross-Cut statistic for a specific parameter.}
\usage{
jc_get_crosscut_graph(
  crosscut_data,
  stats_params,
  colours,
  line_types,
  y_info,
  line_width = 1,
  legend_title = "Statistic"
)
}
\arguments{
\item{crosscut_data}{Data frame with crosscut statistics for the parameter,
as provided by method \code{link{jc_get_crosscut_data}}.}

\item{stats_params}{Vector of stats parameters to plot. See the output for method
\code{link{jc_get_param_stats_table}} to see what stats are available, and
also how the codes (column names) to use. Generally, for percentiles, you can
use 'perc_90', 'perc_25' etc. Also 'mean', 'median', 'max' and 'min.}

\item{colours}{Vector of colour names or HTML colour codes, to match
the statistics codes passed in as parameter \code{stat_codes}.}

\item{line_types}{Vector of line style names to match
the statistics codes passed in as parameter \code{stat_codes} (valid
values are: 'solid', 'dashed', 'dotted', dotdash', 'longdash' etc. ).}

\item{y_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for Y-axis.}

\item{line_width}{Line width for the lines. Try values in range 0.8 to 1.2.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_crosscut_graph} Gets ggplot object showing a crosscut graph that
represents a specific statistic for each budget at a specific epoch.
}
