% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elem_forecast_debug.R
\name{jc_get_elem_forecast}
\alias{jc_get_elem_forecast}
\title{Plot the Area-Under-the-Curve for various Strategies}
\usage{
jc_get_elem_forecast(
  elem_index,
  param_names,
  params_export_file,
  identifiers_to_remove,
  calendar_epochs,
  treatments
)
}
\arguments{
\item{elem_index}{ONE-based index of the element to view}

\item{param_names}{Vector containing the model parameter names to explort.
Typically these will be all parameter names, which you can load from the
'parameters' sheet in the model setup file}

\item{params_export_file}{relative path to the Parameter export/output file}

\item{identifiers_to_remove}{Vector of the identifiers to remove from the
resulting table (e.g. 'area_name', etc. as defined in model lookups under
the 'general' list of the model setup file)}

\item{calendar_epochs}{Vector of calendar epochs for which to show forecast.
You should ensure this vector matches the number of modelling periods exactly.}

\item{treatments}{Data Frame with treatments loaded from the jcass model
treatment exports file}
}
\value{
Data Frame with forecast table for the element
}
\description{
\code{jc_get_elem_forecast} Gets the forecasted values for all parameters for
a specific element identified with its ONE-based element index. That is, to
get the first element in the element set, pass element_index = 1
}
