% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_linear_elements.R
\name{jc_get_element_overlap}
\alias{jc_get_element_overlap}
\title{Get overlap betweeen two linear elements}
\usage{
jc_get_element_overlap(from1, to1, from2, to2)
}
\arguments{
\item{from1}{Start location for the first element}

\item{to1}{End location for the first element}

\item{from2}{Start location for the second element}

\item{to2}{End location for the second element}
}
\value{
length of overlap, in same unit as the location parameters
}
\description{
\code{jc_get_element_overlap} gets the overlap between two linear elements
defined with start locations (from1, to1) and (from2, to2), respectively
}
\examples{
jc_get_element_overlap(0,6,3,10)

}
