% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_analysis.R
\name{jc_get_increm_boxplot}
\alias{jc_get_increm_boxplot}
\title{Gets a Box plot showing increments over epochs for a specific parameter}
\usage{
jc_get_increm_boxplot(increm_data, axis_info, x_label_angle = 0)
}
\arguments{
\item{increm_data}{increment data extracted from parameter export file for
this parameter. This should be the result of method \code{jc_get_increments}.
See \code{link{jc_get_increments}}}

\item{axis_info}{List object containing the following attributes describing
the Y-axis:
\enumerate{
\item 'title' Report-friendly title for the axis
\item 'min' minimum for the axis scale
\item 'max' maximum for the axis scale
\item 'increm' Increment for the axis scale - this will also determine
the bin size if this is a histogram
}}

\item{x_label_angle}{angle for the X-Axis labels (epochs). Default is zero
which shows labels horizontally. To rotate slightly when labels are a bit
cramped, suggest you try -35.}
}
\description{
\code{jc_get_increm_boxplot} Draws a Box Plot showing the distribution of
increments in each epoch for specific parameter.
}
