% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_analysis.R
\name{jc_get_increm_density}
\alias{jc_get_increm_density}
\title{Gets a Density Plot for increments for a specific parameter}
\usage{
jc_get_increm_density(increm_data, x_axis_info)
}
\arguments{
\item{increm_data}{increment data extracted from parameter export file for
this parameter. This should be the result of method \code{jc_get_increments}.
See \code{link{jc_get_increments}}}

\item{x_axis_info}{List object containing the following attributes:
\enumerate{
\item 'title' Report-friendly title for the axis
\item 'min' minimum for the axis scale
\item 'max' maximum for the axis scale
\item 'increm' Increment for the axis scale - this will also determine
the bin size if this is a histogram
}}
}
\description{
\code{jc_get_increm_density} Draws a Density Plot (similar to histogram but
smoothed) of the increments for a specific parameter.
}
