% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increment_analysis.R
\name{jc_get_increments}
\alias{jc_get_increments}
\title{Gets all increments (excluding resets) as a data frame}
\usage{
jc_get_increments(
  params_export_file,
  param_name,
  calendar_epochs,
  expected = "pos"
)
}
\arguments{
\item{params_export_file}{relative path to the jcass parameter exports file}

\item{param_name}{name of the parameter for which to get increments}

\item{calendar_epochs}{calendar epochs for which to get increments.}

\item{expected}{expected slope of increments. Use 'pos' when you expect
positive increments such as for rut depth (which increases over time), and
'neg' when you expect negative increments such as for texture depth (which
decreases over time).}
}
\value{
Vector of increments in which the increment for calendar epoch X is
the increment from X-1 to X. For example, the increment shown for epoch
2025 will be the increment from 2024 to 2025
}
\description{
\code{jc_get_increments} parses the increments from a parameter export file
and returns the information in a dataset for further graphics and statistical
analysis.
}
\details{
Any resets, in which the increment is opposite to what is expected, are
assigned NA values and thus will be excluded from downstream statistics and
graphics analysis.
}
