% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{jc_get_lookups}
\alias{jc_get_lookups}
\title{Set's up a dictionary of Lookup tables for fast lookup of constants or
control parameters}
\usage{
jc_get_lookups(lkp_url, is_local = TRUE)
}
\arguments{
\item{lkp_url}{URL for the Excel file containing the lookup sheets}

\item{is_local}{Flag to indicate if the source URL maps to a local file (TRUE),
or a file on the web}
}
\value{
\code{list} object with keys mapping to lookup set names, and values
containing data frames that can be lookup up on
}
\description{
This method reads a set of lookups from Excel and builds a \code{list} object
with keys/attribute names matching the sheet names. Lookups are only read
from sheets starting with 'lkp_'.
}
\details{
Each lookup key (sheet name) in the returned object maps to a dataframe in
which the row names match the values read from the left-most column. Values
in other columns can then be lookup up (assuming exact matches) in any of the
other columns.

The method getlookupvalue() on the model ('jc_model') object can be used to
conveniently find the value mapping to a specific lookup set, key and column.
}
