% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{jc_get_NA_percentages}
\alias{jc_get_NA_percentages}
\title{Utility function to get number and percentage of NAs in each column}
\usage{
jc_get_NA_percentages(df, cols_to_check = NULL)
}
\arguments{
\item{df}{data frame to evaluate}

\item{cols_to_check}{columns to check or pass in NULL to check all columns}
}
\value{
data frame with columns:
\enumerate{
\item 'column' containing the column name
\item 'na_count' containing the number of NA values in each column
\item 'na_percent' containing the percentage of NA values in each column
}
}
\description{
\code{jc_get_NA_percentages} gets the number of NA values in each column and
also converts this to a percentage of NA based on the total number of rows in
df
}
