% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_breakdown_info.R
\name{jc_get_network_breakdown}
\alias{jc_get_network_breakdown}
\title{Gets a breakdown table based on a specific grouping column}
\usage{
jc_get_network_breakdown(
  raw_data,
  group_col,
  loc_from_col,
  loc_to_col,
  scale_fact = 1000
)
}
\arguments{
\item{raw_data}{raw data (as data frame) from which to calculate lengths.}

\item{group_col}{columns on which to get length grouping. Assumed to be a
factor or character column.}

\item{loc_from_col}{Name of the column that contains the 'from location' or
'start' of each element}

\item{loc_to_col}{Name of the column that contains the 'to location' or
'end' of each element}

\item{scale_fact}{Scaling factor for converting lengths. E.g. pass 1000 to
express the lengths in KM when the original lengths are in Metres.}
}
\description{
\code{jc_get_network_breakdown} Gets a breakdown table based on a specific
column that shows the length and percentage length for each group.
}
