% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_breakdown_info.R
\name{jc_get_network_breakdown_nice}
\alias{jc_get_network_breakdown_nice}
\title{Gets a nicely formatted breakdown table based on a specific grouping column}
\usage{
jc_get_network_breakdown_nice(
  raw_data,
  group_col,
  group_col_label,
  length_label,
  loc_from_col,
  loc_to_col,
  scale_fact = 1000,
  length_decimals = 2,
  percent_decimals = 2
)
}
\arguments{
\item{raw_data}{raw data (as data frame) from which to calculate lengths.}

\item{group_col}{columns on which to get length grouping. Assumed to be a
factor or character column.}

\item{group_col_label}{User/Report-friendly label for the grouping column.}

\item{length_label}{User/Report-friendly label for the length column.}

\item{loc_from_col}{Name of the column that contains the 'from location' or
'start' of each element}

\item{loc_to_col}{Name of the column that contains the 'to location' or
'end' of each element}

\item{scale_fact}{Scaling factor for converting lengths. E.g. pass 1000 to
express the lengths in KM when the original lengths are in Metres.}

\item{length_decimals}{Number of decimals to show on the length column.}

\item{percent_decimals}{Number of decimals to show on the percentage length
column.}
}
\description{
\code{jc_get_network_breakdown_nice} Gets a gt breakdown table based on a specific
column that shows the length and percentage length for each group.
}
