% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_param_plots.R
\name{jc_get_param_stats_plot}
\alias{jc_get_param_stats_plot}
\title{Gets a graph showing stats for a parameter over epochs.}
\usage{
jc_get_param_stats_plot(
  param_stats,
  stat_codes,
  colours,
  line_types,
  x_info,
  y_info,
  line_width = 1,
  legend_title = "Statistic"
)
}
\arguments{
\item{param_stats}{Data frame with statistics for the parameter, as provided
by method \code{link{jc_get_param_stats_table}}.}

\item{stat_codes}{Vector of stats parameters to plot. See the output for method
\code{link{jc_get_param_stats_table}} to see what stats are available, and
also how the codes (column names) to use. Generally, for percentiles, you can
use 'perc_90', 'perc_25' etc. Also 'mean', 'median', 'max' and 'min.}

\item{colours}{Vector of colour names or HTML colour codes, to match
the statistics codes passed in as parameter \code{stat_codes}.}

\item{line_types}{Vector of line style names to match
the statistics codes passed in as parameter \code{stat_codes} (valid
values are: 'solid', 'dashed', 'dotted', dotdash', 'longdash' etc. ).}

\item{x_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for X-axis.}

\item{y_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for Y-axis.}

\item{line_width}{Line width for the lines. Try values in range 0.8 to 1.2.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_param_stats_plot} Gets ggplot object showing a lines representing
various statistics for a specific parameter over the
modelling epochs.
}
\details{
Also see \code{link{jc_get_spend_summary_multi_budget}}
}
