% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_plots_multirun.R
\name{jc_get_param_stats_plot_multirun}
\alias{jc_get_param_stats_plot_multirun}
\title{Gets a graph showing stats for a parameter for multiple scenarios.}
\usage{
jc_get_param_stats_plot_multirun(
  model_param,
  stat_params,
  budget_codes,
  colours,
  line_types,
  x_info,
  y_info,
  calendar_epochs,
  line_width = 1,
  outputs_base = "outputs/paramdata_",
  legend_title = "Budget Level"
)
}
\arguments{
\item{model_param}{Name of the model parameter to get the graph for.}

\item{stat_params}{Vector of stats parameters to plot. See the output for method
\code{link{jc_get_param_stats_table}} to see what stats are available, and
also how the codes (column names) to use. Generally, for percentiles, you can
use 'perc_90', 'perc_25' etc. Also 'mean', 'median', 'max' and 'min.}

\item{budget_codes}{Vector of run/budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}

\item{colours}{Vector of colour names or HTML colour codes, to match budget
codes}

\item{line_types}{Vector of line style names to match statistics parameters
(valid values are: 'solid', 'dashed', 'dotted', dotdash', 'longdash' etc. )}

\item{x_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for X-axis.}

\item{y_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for Y-axis.}

\item{calendar_epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{line_width}{Line width for the lines. Try values in range 0.8 to 1.2.}

\item{outputs_base}{Base stub for relative path to parameter output files.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_param_stats_plot_multirun} Gets ggplot object showing a line
graph with lines representing the statistics for a specific parameter over the
modelling epochs and over all the modelling runs in run_codes. Colours are
used to define budget scenario, and line style is used to define statistic.
}
\details{
Also see \code{link{jc_get_spend_summary_multi_budget}}
}
