% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_param_stats.R
\name{jc_get_param_stats_table}
\alias{jc_get_param_stats_table}
\title{Gets a table showing statistics for a parameter.}
\usage{
jc_get_param_stats_table(param_data, epochs)
}
\arguments{
\item{param_data}{Data frame containing the data for the model parameter as
loaded from the jcass output file.}

\item{epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}
}
\description{
\code{jc_get_param_stats_table_multirun} Gets a table showing the statistics
for a specific model and statistics parameter.
}
