% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_stats_multirun.R
\name{jc_get_param_stats_table_multirun}
\alias{jc_get_param_stats_table_multirun}
\title{Gets a table showing statistics for a parameter for multiple scenarios.}
\usage{
jc_get_param_stats_table_multirun(
  model_param,
  stat_param,
  run_codes,
  calendar_epochs,
  outputs_base = "outputs/paramdata_"
)
}
\arguments{
\item{model_param}{Name of the model parameter to get
See \code{link{jc_get_spend_summary_multi_budget}}.}

\item{stat_param}{Statistics parameter to get data for.}

\item{run_codes}{Vector of run codes. It is assumed that a matching parameter
output file (with run code appended) are in the outputs/ subfolder of the
working folder.}

\item{calendar_epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{outputs_base}{Base stub for relative path to parameter output files.}
}
\description{
\code{jc_get_param_stats_table_multirun} Gets a table showing the statistics
for a specific model and statistics parameter.
}
