% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_param_stats.R
\name{jc_get_param_threshold_lengths}
\alias{jc_get_param_threshold_lengths}
\title{Gets a table showing above and below threshold for a parameter.}
\usage{
jc_get_param_threshold_lengths(param_data, epochs, threshold, from_col, to_col)
}
\arguments{
\item{param_data}{Data frame containing the data for the model parameter as
loaded from the jcass output file.}

\item{epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{threshold}{threshold to evaluate}

\item{from_col}{name of the from/start column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}

\item{to_col}{name of the to/end column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}
}
\description{
\code{jc_get_param_threshold_lengths} Gets a table showing the lengths and
percentages above and below a given threshold for a specific model parameter.
}
\details{
Note: this function is relevant for elements in linear networks (e.g. road
and water networks). For domains that require counts instead of lengths, please
notify Lonrix and we can add an extension to this method that gives counts
instead of lengths.
}
