% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_plots_multirun.R
\name{jc_get_param_threshold_plot_multirun}
\alias{jc_get_param_threshold_plot_multirun}
\title{Gets a graph showing stats for a parameter for multiple scenarios.}
\usage{
jc_get_param_threshold_plot_multirun(
  model_param,
  thresholds,
  from_col,
  to_col,
  show_percent,
  show_below,
  budget_codes,
  colours,
  line_types,
  x_info,
  y_info,
  calendar_epochs,
  line_width = 1,
  outputs_base = "outputs/paramdata_",
  legend_title = "Budget Level"
)
}
\arguments{
\item{model_param}{Name of the model parameter to get the graph for.}

\item{thresholds}{Vector of thresholds to evaluate.}

\item{from_col}{name of the from/start column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}

\item{to_col}{name of the to/end column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}

\item{show_percent}{Pass TRUE to show 'percentage length' or FALSE to show
absolute length based on specified from and to column names}

\item{show_below}{Pass TRUE to show 'less than or equal' to threshold or
FALSE to show 'greater than' values.}

\item{budget_codes}{Vector of run/budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}

\item{colours}{Vector of colour names or HTML colour codes, to match budget
codes}

\item{line_types}{Vector of line style names to match statistics parameters
(valid values are: 'solid', 'dashed', 'dotted', dotdash', 'longdash' etc. )}

\item{x_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for X-axis.}

\item{y_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for Y-axis.}

\item{calendar_epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{line_width}{Line width for the lines. Try values in range 0.8 to 1.2.}

\item{outputs_base}{Base stub for relative path to parameter output files.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_param_stats_plot_multirun} Gets ggplot object showing a line
graph with lines representing the statistics for a specific parameter over the
modelling epochs and over all the modelling runs in run_codes. Colours are
used to define budget scenario, and line style is used to define statistic.
}
\details{
Also see \code{link{jc_get_spend_summary_multi_budget}}
}
