% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_stats_multirun.R
\name{jc_get_param_threshold_table_multirun}
\alias{jc_get_param_threshold_table_multirun}
\title{Gets a table showing above and below threshold for multiple scenarios.}
\usage{
jc_get_param_threshold_table_multirun(
  model_param,
  run_codes,
  calendar_epochs,
  threshold,
  from_col,
  to_col,
  outputs_base = "outputs/paramdata_"
)
}
\arguments{
\item{model_param}{Name of the model parameter to get
See \code{link{jc_get_spend_summary_multi_budget}}.}

\item{run_codes}{Vector of run codes. It is assumed that a matching parameter
output file (with run code appended) are in the outputs/ subfolder of the
working folder.}

\item{calendar_epochs}{Vector with the calendar epochs for which to get stats. The numbers
in this vector must match the columns in the parameter output file, otherwise
errors will occur.}

\item{threshold}{threshold to evaluate}

\item{from_col}{name of the from/start column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}

\item{to_col}{name of the to/end column for element lengths. This column
must exist in your model output data (specified in model setup as an identifier
column)}

\item{outputs_base}{Base stub for relative path to parameter output files.}
}
\description{
\code{jc_get_param_threshold_table_multirun} Gets a table showing the length
and percentage lengths above and below a given threshold for a parameter.
}
\details{
Note: this function is relevant for elements in linear networks (e.g. road
and water networks). For domains that require counts instead of lengths, please
notify Lonrix and we can add an extension to this method that gives counts
instead of lengths.
}
