% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_setup.R
\name{jc_get_path_setup}
\alias{jc_get_path_setup}
\title{Sets up a List object with file paths for pre-and-postprocessing}
\usage{
jc_get_path_setup(paths_setup_file)
}
\arguments{
\item{paths_setup_file}{Excel file with path setup information for jcass. It
is assumed this file has a sheet named 'paths' and that this sheet is in the
proper template format}
}
\description{
\code{jc_get_path_setup} Reads path information from a setup file, then does
various checks and creates the necessary folders for pre-processing,
calibration and  running a Juno Cassandra model, and post-processing results.
}
