% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{jc_get_PWF}
\alias{jc_get_PWF}
\title{Gets the Present Worth Factor (PWF)}
\usage{
jc_get_PWF(period, inflation_percent, discount_percent)
}
\arguments{
\item{period}{number of periods over which to discount and inflate}

\item{inflation_percent}{Inflation rate as a percentage (i.e. 4.5, not 0.045)}

\item{discount_percent}{Discount rate as a percentage (i.e. 4.5, not 0.045)}
}
\value{
real number which is the Present Worth Factor. Multiply cost by this
factor to get the present value of the amount
}
\description{
\code{jc_debug_elem} Gets the Present Worth Factor (PWF) for a given
number of periods, a given inflation rate and a given discount rate
}
