% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{jc_get_rankpercent}
\alias{jc_get_rankpercent}
\title{Gets a Percent Rank based on data in a column of a data frame}
\usage{
jc_get_rankpercent(df, column_name, reverse = FALSE)
}
\arguments{
\item{df}{data frame with column on which to rank}

\item{column_name}{name of the column on which to get percent ranking}

\item{reverse}{If TRUE, the smallest value will have a rank percent of 100}
}
\value{
vector with percent rank values, which will have a maximum of 100, but
minimum may not be zero.
}
\description{
\code{jc_get_rankpercent} does a rank percent calculation on a column in a
data frame.
}
\details{
When parameter \code{reverse} is FALSE (the default), the largest absolute
value will have a percentage rank of 100. Note that the minimum value will
not necessarily have a corresponding rank percent value of zero.

When parameter \code{reverse} is TRUE, the smallest absolute
value will have a percentage rank of 100. Note that the maximum value will
not necessarily have a corresponding rank percent value of zero.
}
