% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_setup.R
\name{jc_get_sheets_data}
\alias{jc_get_sheets_data}
\title{Loads data on sheets in an Excel file into a List of data frames}
\usage{
jc_get_sheets_data(lkp_url, sheet_prefix, is_local = TRUE)
}
\arguments{
\item{lkp_url}{URL for the Excel file containing the lookup sheets}

\item{sheet_prefix}{Prefix that identifies sheets to read data from}

\item{is_local}{Flag to indicate if the source URL maps to a local file (TRUE),
or a file on the web}
}
\value{
\code{list} object with keys mapping to lookup set names, and values
containing data frames
}
\description{
This method reads a set of data frames from Excel and builds a \code{list}
object in which the keys (attribute names) match the sheet names. The sheets
from which to read are identified by a specified prefix.
}
\details{
For example, to read the data from all sheets starting with "model_", you
should provide the prefix 'model_" as a parameter. The data on each sheet
should be a monolithic block of data with top-left corner in cell A1, and
with valid column names#'
}
