% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_proc_spend_plots.R
\name{jc_get_spend_graph}
\alias{jc_get_spend_graph}
\title{Gets a graph showing percentage Budget Utilisation over time.}
\usage{
jc_get_spend_graph(
  spending_info,
  categories,
  y_info,
  colours,
  line_types,
  line_width = 1,
  legend_title = "Category"
)
}
\arguments{
\item{spending_info}{Spending info retrieved with
\code{link{jc_get_spending_info}}}

\item{categories}{Vector of category names}

\item{y_info}{List object containing attributes 'min', 'max', 'increm' and
'title' for Y-axis.}

\item{colours}{Vector of colour names to match category names}

\item{line_types}{Vector of line style names to match category names
the statistics codes passed in as parameter \code{stat_codes} (valid
values are: 'solid', 'dashed', 'dotted', dotdash', 'longdash' etc. ).}

\item{line_width}{Line width for the lines. Try values in range 0.8 to 1.2.}

\item{legend_title}{Title for the legend.}
}
\description{
\code{jc_get_spend_graph} Gets ggplot object showing lines representing
budget utilisation for different treatment categories over time.
}
