% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spend_summaries.R
\name{jc_get_spend_info_multi_budget}
\alias{jc_get_spend_info_multi_budget}
\title{Gets a table summarising spending across several budget codes.}
\usage{
jc_get_spend_info_multi_budget(
  budget_codes,
  budgets_file,
  network_length,
  treats_file_base = "outputs/treatments_"
)
}
\arguments{
\item{budget_codes}{vector of budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}

\item{budgets_file}{Relative path to the Excel file containing the budgets.}

\item{network_length}{Network length from which the percentage of length
treatment etc. can be calculated}

\item{treats_file_base}{Relative path to the Excel file containing the
treatments exported from jcass. This will not be the full file name, but only
the stub from which the full name can be built by appending the budget code
at the end.}
}
\description{
\code{jc_get_spend_info_multi_budget} Gets a table that summarises spending across
different budget codes. The result contains combined data for all budget
codes, and for each budget code provides summary data relating to the
following parameters:
\enumerate{
\item 'project_count' Report-friendly title for the axis
\item 'min' minimum for the axis scale
\item 'max' maximum for the axis scale
\item 'increm' Increment for the axis scale - this will also determine
the bin size if this is a histogram
}
}
\details{
This method is similar to \code{jc_get_spend_param_multi_budget}. However,
this method returns the summaries for all spending parameters
whereas method \code{jc_get_spend_param_multi_budget} returns the data only for one
specific spending parameter. See \code{link{jc_get_spend_param_multi_budget}}
}
