% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spend_summaries.R
\name{jc_get_spend_param_multi_budget}
\alias{jc_get_spend_param_multi_budget}
\title{Gets a table summarising a spending parameter across several budget codes.}
\usage{
jc_get_spend_param_multi_budget(
  budget_codes,
  budgets_file,
  spend_param,
  network_length,
  min_calendar_period,
  max_calendar_period,
  treats_file_base = "outputs/treatments_"
)
}
\arguments{
\item{budget_codes}{vector of budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}

\item{budgets_file}{Relative path to the Excel file containing the budgets.}

\item{spend_param}{spending parameter for which to get information. Valid
values are (case sensitive!):
\enumerate{
\item 'project_count' mean number of projects for category across all periods
\item 'avg_spend' mean spending (in millions) for category across all periods
\item 'total_spend' sum of all spending (in millions) for category across all periods
\item 'avg_perc' mean of percentage of network treated for category
across all periods
\item 'total_perc' sum of annual percentages of network treated for category
across all periods
\item 'avg_length' mean of length treated for category across all periods
\item 'total_length' sum of length treated for category across all periods
\item 'budget_util' Average for percentage budget utilised for category
across all periods
}}

\item{network_length}{Network length from which the percentage of length
treatment etc. can be calculated.}

\item{min_calendar_period}{minimum treatment period (in calendar terms) to
consider (inclusive).}

\item{max_calendar_period}{maximum treatment period (in calendar terms) to
consider (inclusive).}

\item{treats_file_base}{Relative path to the Excel file containing the
treatments exported from jcass. This will not be the full file name, but only
the stub from which the full name can be built by appending the budget code
at the end.}
}
\description{
\code{jc_get_spend_param_multi_budget} Gets a table that summarises spending across
different budget codes for a specific spending parameter (e.g. project count).
The result contains combined data for all budget codes, and for each budget
code provides summary data relating to the specified parameter.
}
\details{
This method is similar to \code{link{jc_get_spend_info_multi_budget}}.
However, this method returns the summaries for all spending parameters
whereas method \code{jc_get_spend_param_multi_budget} returns the data only for one
specific spending parameter. See \code{link{jc_get_spend_info_multi_budget}}.
}
