% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spend_summaries.R
\name{jc_get_spend_summary_multi_budget}
\alias{jc_get_spend_summary_multi_budget}
\title{Gets a table summarising spending for over multiple budgets.}
\usage{
jc_get_spend_summary_multi_budget(
  budget_codes,
  budgets_file,
  network_length,
  min_calendar_period,
  max_calendar_period,
  spend_params = c("avg_spend", "total_spend", "avg_perc", "total_perc", "avg_length",
    "total_length"),
  treats_file_base = "outputs/treatments_"
)
}
\arguments{
\item{budget_codes}{vector of budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}

\item{budgets_file}{Relative path to the Excel file containing the budgets.}

\item{network_length}{Network length from which the percentage of length
treatment etc. can be calculated}

\item{min_calendar_period}{minimum treatment period (in calendar terms) to
consider (inclusive).}

\item{max_calendar_period}{maximum treatment period (in calendar terms) to
consider (inclusive).}

\item{spend_params}{Vector of spending parameter codes for which to get the
summary. Valid values are:
\enumerate{
\item 'project_count' mean number of projects for category across all periods
\item 'avg_spend' mean spending (in millions) for category across all periods
\item 'total_spend' sum of all spending (in millions) for category across all periods
\item 'avg_perc' mean of percentage of network treated for category
across all periods
\item 'total_perc' sum of annual percentages of network treated for category
across all periods
\item 'avg_length' mean of length treated for category across all periods
\item 'total_length' sum of length treated for category across all periods
\item 'budget_util' Average for percentage budget utilised for category
across all periods
}}

\item{treats_file_base}{Relative path to the Excel file containing the
treatments exported from jcass. This will not be the full file name, but only
the stub from which the full name can be built by appending the budget code
at the end.}
}
\description{
\code{jc_get_spend_summary_multi_budget} Gets a table that summarises spending
grouped by treatment category. Provides summary data relating to a chosen set
of parameters:
}
