% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spend_summaries.R
\name{jc_get_spend_summry_multi_budget_gt}
\alias{jc_get_spend_summry_multi_budget_gt}
\title{Gets a report ready table summarising spending for over multiple budgets.}
\usage{
jc_get_spend_summry_multi_budget_gt(summary_table, budget_codes)
}
\arguments{
\item{summary_table}{data frame containing the output from method
\code{jc_get_spend_summary_multi_budget}.
See \code{link{jc_get_spend_summary_multi_budget}}}

\item{budget_codes}{vector of budget codes, e.g. 'upper', 'lower', 'base' etc.
It is assumed that there will be output files matching these codes and also
that the budgets file will have matching sheets containing the budgets for
each of these codes.}
}
\description{
\code{jc_get_spend_summry_multi_budget_gt} Gets a report ready gt() table
that summarises spending grouped by treatment category. This method takes
the output of method \code{jc_get_spend_summary_multi_budget} and just
converts it to a printed image.
}
\details{
Also see \code{link{jc_get_spend_summary_multi_budget}}
}
