% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spend_summaries.R
\name{jc_get_spending_info}
\alias{jc_get_spending_info}
\title{Gets a table summarising spending for a specific budget.}
\usage{
jc_get_spending_info(treatments, budget, network_length)
}
\arguments{
\item{treatments}{Data frame containing the treatments list exported by
jcass.}

\item{budget}{Data frame containing the budget.}

\item{network_length}{Network length from which the percentage of length
treatment etc. can be calculated.}
}
\description{
\code{jc_get_spending_info} Gets a table that summarises spending grouped by
treatment category and period for a specific budget. Provides summary data
relating to the following parameters:
\enumerate{
\item 'number' number of projects
\item 'total_cost' total cost for category and period
\item 'total_length' total treatment length for category and period
\item 'total_perc' percentage of total length treatment for category
and period
\item 'budget_util' Percentage of budget utilised for category and period
}
}
