% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve_helper.R
\name{jc_get_strategy_table}
\alias{jc_get_strategy_table}
\title{Gets a table with strategy summary information for various Strategies}
\usage{
jc_get_strategy_table(strat_summary_data, strat_details)
}
\arguments{
\item{strat_summary_data}{data frame with data loaded from the 'summary'
sheet of the Strategy Debug export file}

\item{strat_details}{data frame with data loaded from the 'detail'
sheet of the Strategy Debug export file}
}
\value{
Data Frame that can be further customised or used for graphs etc
}
\description{
\code{jc_get_strategy_table} Gets a nice table with summary information for
each treatment strategy using data from a jcass Strategy Debug export file.
}
