% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_methods1.R
\name{jc_get_subset}
\alias{jc_get_subset}
\title{Creates a subset that retains only a certain number of columns}
\usage{
jc_get_subset(data, columns)
}
\arguments{
\item{data}{Data frame with source data on which the columns will be checked}

\item{columns}{Character vector containing the list of columns to retain. If}
}
\value{
data frame with only the required columns retained
}
\description{
\code{jc_get_subset} returns a subset of the input data that contains only
the specified number of columns
}
