% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budget_helpers.R
\name{jc_get_unlimited_budget}
\alias{jc_get_unlimited_budget}
\title{Creates an Unlimited Budget suitable for a specified model}
\usage{
jc_get_unlimited_budget(model)
}
\arguments{
\item{model}{\code{jm_model} model object correctly set up}
}
\value{
a data frame containing budget categories in columns, and a matching
amount for each modelling period in rows.
}
\description{
\code{jc_get_unlimited_budget} Creates an unlimited budget with automatically
added budget categories based on the treatment categories defined in your
model. The number of budget periods added are also automatically set to
match the number of modelling periods
}
