% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_moora.R
\name{jc_moora_rank}
\alias{jc_moora_rank}
\title{Multi-Objective Optimisation on the basis of Ratio Analysis (MOORA)}
\usage{
jc_moora_rank(data, weights, obj_types)
}
\arguments{
\item{data}{alternatives in rows and criteria in columns. It is assumed that
all columns are objectives}

\item{weights}{unit weights for each objective (must add up to one)}

\item{obj_types}{Objective types for each objective. Must match the number
of columns in data. Use 'max' to indicate an objective to maximize, and 'min'
to indicate an objective to minimize (e.g. cost)}
}
\description{
Calculations and unit testing based on this video:
}
\details{
Rank (1 is best) and the normalised value used to rank is added to the
data set which is then returned
}
