% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{jc_normalise_min_max}
\alias{jc_normalise_min_max}
\title{Converts values to Min-Max Normalised values}
\usage{
jc_normalise_min_max(values, min_val = NA, max_val = NA)
}
\arguments{
\item{values}{numerical vector of values to normalise}

\item{min_val}{optional custom minimum value to use. If ommitted, the minimum
used for normalisation will be automatically calculated from \code{values}}

\item{max_val}{optional custom maximum value to use. If ommitted, the maximum
used for normalisation will be automatically calculated from \code{values}}
}
\value{
vector of normalised values
}
\description{
\code{jc_normalise_min_max} takes a vector of values and returns a matching
min-max normalised set. If min and max values are not explicitly supplied,
the minimum and maximum values will be automatically calculated from the
values passed
}
\details{
For more information on min-max normalisation of values, see
\url{https://en.wikipedia.org/wiki/Feature_scaling}
}
\examples{
values <- c(0,2,4,6,8,10) #vector of values to normalise
#for this example, the mean is 5 and the standard deviation is 3.7417
norm_values <- jc_normalise_min_max(values)
norm_values  #Should be 0, 0.2, 0.4, 0.6, 0.8, 1

# Example with scalar values and custom min and max
# 65 lies halfway between 10 and 120. Should get 0.5
jc_normalise_min_max(65, 10, 120)
}
