% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{jc_normalise_percentiles}
\alias{jc_normalise_percentiles}
\title{Converts values to Percentile-Based Scaled Normalised values}
\usage{
jc_normalise_percentiles(values, quantiles, norm_scales)
}
\arguments{
\item{values}{numerical vector of values to normalise}

\item{quantiles}{vector of quantiles to match with the scales (must contain
values between 0 and 1)}

\item{norm_scales}{scaled values that represent the normalised range of
values to match each specified quantile. Size of this vector must match the
size of the \code{quantiles} parameter.}
}
\description{
\code{jc_normalise_percentiles} takes a vector of values and returns a matching
normalised set. In this approach, a scale is first built that matches
quantiles with specified scaled values. For example, for quantile 0.5 you may
have a matching scaled value of zero, and for quantile 1.0 (maximum) you may
have a scaled value of 10. A piecewise linear model is then use to
interpolate between the quantile values and the scaled values.
}
