% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_engineering.R
\name{jc_normalise_z}
\alias{jc_normalise_z}
\title{Converts values to Z-score Normalised values}
\usage{
jc_normalise_z(values, mean_val = NA, stdev = NA)
}
\arguments{
\item{values}{numerical vector of values to normalise}

\item{mean_val}{optional custom mean value to use. If ommitted, the mean
used for normalisation will be automatically calculated from \code{values}}

\item{stdev}{optional custom standard deviation to use. If ommitted, the
standard deviation used for normalisation will be automatically calculated
from \code{values}}
}
\value{
vector of normalised values
}
\description{
\code{jc_normalise_z} takes a vector of values and returns a matching
set of Z-score normalised values. The Z-score is defined as:
\code{(value - mean)/std_dev}. If the mean and standard deviation values are
not explicitly supplied, they will be automatically calculated from the
values passed
}
\details{
Note: If standard deviation is NA or zero, then a score of 0 is returned for
all values. This is because, if there is no standard deviation, then all
values are the same which will be the mean. In that case, all z-scores are
zero.

For more information on Z-Score normalisation of values, see
\url{https://en.wikipedia.org/wiki/Feature_scaling}
}
\examples{
values <- c(0,2,4,6,8,10) #vector of values to normalise
#for this example, the mean is 5 and the standard deviation is 3.7417
norm_values <- jc_normalise_z(values)
#Should be -1.3363,	-0.8018,	-0.2673,	0.2673,	0.8018,	1.3363
}
