% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve_helper.R
\name{jc_plot_strategy_table_gt}
\alias{jc_plot_strategy_table_gt}
\title{Gets a print-ready Table with strategy summary information for various Strategies}
\usage{
jc_plot_strategy_table_gt(header, strat_summary_data, strat_details)
}
\arguments{
\item{header}{header for the table}

\item{strat_summary_data}{data frame with data loaded from the 'summary'
sheet of the Strategy Debug export file}

\item{strat_details}{data frame with data loaded from the 'detail'
sheet of the Strategy Debug export file}
}
\value{
GT package table object
}
\description{
\code{jc_plot_strategy_table_gt} Gets a nice printable table with summary
information for each treatment strategy using data from a jcass Strategy Debug export file.
}
