% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_methods1.R
\name{jc_type_columns}
\alias{jc_type_columns}
\title{Converts columns to specified types}
\usage{
jc_type_columns(data, columns)
}
\arguments{
\item{data}{Data frame with raw input data in which the columns will
be converted}

\item{columns}{Data frame containing information on the columns to
be converted. This dataframe must contain column names: 'column_name' and
'data_type'. Column names specified here must match (case sensitive) the
column names in the raw input data otherwise an error will be thrown}
}
\value{
A data frame identical to the source data but with columns specified
in the \code{columns} parameter converted to the specified types.
}
\description{
\code{jc_type_columns} Loops over a set of specified columns, each with an
accompanying data type specifier. Columns in the raw data set are then
converted to the specified type. Any columns in the raw input data that are
not listed in the columns parameter are left untouched.
}
\details{
For numeric columns, and any text values found in the column will be
automatically converted to NA.
}
