% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelling.R
\name{jc_update_gen_param}
\alias{jc_update_gen_param}
\title{Updates a specific setting in a General Parameters data frame}
\usage{
jc_update_gen_param(gen_params, param_code, value)
}
\arguments{
\item{gen_params}{Data frame containing General Modelling parameters with
columns 'setting_name' and 'value'}

\item{param_code}{The code for the parameter to update. See online help for
the valid parameter codes for General Parameters (case sensitive!)}

\item{value}{The new value to set for the specified parameter}
}
\description{
\code{jc_update_gen_param} Is a utility function to quickly update
a specific setting (e.g. discount rate) in a data frame containing
General Model parameters.
}
