% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_modelling.R
\name{jc_update_num_mod_periods}
\alias{jc_update_num_mod_periods}
\title{Updates the number of modelling periods in General Parameters}
\usage{
jc_update_num_mod_periods(gen_params, num_periods)
}
\arguments{
\item{gen_params}{Data frame containing General Modelling parameters with
columns 'setting_name' and 'value'}

\item{num_periods}{Number of modelling periods}
}
\description{
\code{jc_update_num_mod_periods} Is a utility function to quickly update
the value associated with the parameter 'number_of_modelling_periods' in
a data frame containing General Model parameters.
}
