% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_utils.R
\name{jc_update_setup_parameter}
\alias{jc_update_setup_parameter}
\title{Updates a parameter in a jcass setup file}
\usage{
jc_update_setup_parameter(
  setup_file,
  sheet_name,
  setting_key,
  setting_value,
  key_column = "setting_name",
  value_column = "value"
)
}
\arguments{
\item{setup_file}{relative path to the setup file to update}

\item{sheet_name}{name of the sheet that holds the data}

\item{setting_key}{unique KEY that identifies the setting value to be updated}

\item{setting_value}{new value to set for this setting}

\item{key_column}{name of the column that contains the key on which to look
up the row}

\item{value_column}{name of the column that holds the value to be updated.}
}
\description{
\code{jc_update_setup_parameter} Is a utility function that allows you to
update a parameter in an Excel setup file for jcass without having to open
the file in Excel. Ensure the file is NOT already open!
}
\details{
This function allows you to dymically update parameters for example when
you want to run a model in a loop with different treshold parameters.

IMPORTANT! Be careful when using this function to update a parameter in your
lookup lists. This function will search for the FIRST key value that matches
the value passed as 'setting_key'. If the values in the 'key_column' are not
unique, then only the first occurence's value will be updated.
}
