% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{jc_utils_get_plmvalue}
\alias{jc_utils_get_plmvalue}
\title{Do PieceWise Linear Interpolation and return result.}
\usage{
jc_utils_get_plmvalue(x, xpoints, ypoints, extrapolate = FALSE)
}
\arguments{
\item{x}{x coordinate on which to get the function value}

\item{xpoints}{x coordinates that define the piecewise linear model. Assumed
to be in order of increasing values.}

\item{ypoints}{y coordinates that define the piecewise linear model}

\item{extrapolate}{TRUE or FALSE value indicating whether to extrapolate for
x-values below or above the limits in xpoints}
}
\value{
real value representing the interpolated or extrapolated value
}
\description{
\code{jc_utils_get_plmvalue} does an interpolation on a piecewise linear
model. The function can optionally constrain the results at the outer limits
or extrapolate.
}
