% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm.R
\name{jc_utils_get_plmvalues}
\alias{jc_utils_get_plmvalues}
\title{Vectorised version to Do PieceWise Linear Interpolation and return results.}
\usage{
jc_utils_get_plmvalues(xvalues, xpoints, ypoints, extrapolate = FALSE)
}
\arguments{
\item{xvalues}{vector of x coordinates on which to get the function values}

\item{xpoints}{x coordinates that define the piecewise linear model. Assumed
to be in order of increasing values.}

\item{ypoints}{y coordinates that define the piecewise linear model}

\item{extrapolate}{TRUE or FALSE value indicating whether to extrapolate for
x-values below or above the limits in xpoints}
}
\value{
vector of real values representing the interpolated or extrapolated
values
}
\description{
\code{jc_utils_get_plmvalues} does an interpolation on a piecewise linear
model. The function can optionally constrain the results at the outer limits
or extrapolate.
}
